+++ docs/example/regionDownload.md
@@ -1,0 +1,14 @@
+---
+toc: false
+order: 100
+nav:
+  title: 示例
+  path: /example
+  order: 3
+---
+
+---
+
+## 行政区域下载
+
+<code src="./regionDownload/index.tsx" compact defaultShowCode></code>
