--- src/components/RegionDownload/index.less
@@ -1,23 +1,0 @@
-.panel {
-  position: absolute;
-  right: 0;
-  z-index: 1000;
-  width: 300px;
-  height: 100%;
-  padding: 10px 20px;
-  background: #fff;
-
-  .source-select {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-  
-  .download-content {
-    margin-top: 10px;
-    .data-input {
-      display: flex;
-      margin-top: 5px;
-    }
-  }
-}
