--- src/components/RegionDownload/index.md
@@ -1,25 +1,0 @@
----
-toc: content
-order: 3
-group:
-  title: 行政区域下载
-  order: 100
-nav:
-  title: 组件
-  path: /components
-hide: true
----
-
-## 行政区域下载 - RegionDownload
-
-### 介绍
-
-### 使用场景
-
-### 代码演示
-
-#### 默认示例
-
-<code src="./demos/default.tsx" compact defaultShowCode></code>
-
-<API></API>
